feat: AuthContext, NotificationContext 관련 변수, 함수 관심사 분리 완료
