+++ src/@lib/hooks/useNotificationContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { NotificationContext } from "../../context/NotificationContext";
+import { NotificationContext } from "../../context/NotificationContext.tsx";
 
 export const useNotificationContext = () => {
   const context = useContext(NotificationContext);
