--- src/context/AuthContext.ts
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-import { AuthContextType } from "../type/type";
-
-export const AuthContext = createContext<AuthContextType | undefined>(
-  undefined
-);
