+++ src/context/AuthContext.tsx
@@ -1,0 +1,39 @@
+import { createContext, useState } from "react";
+import { useCallback, useMemo } from "../@lib";
+import { AuthContextType, ContextProps, User } from "../type/type";
+import { useNotificationContext } from "../@lib/hooks/useNotificationContext";
+
+export const AuthContext = createContext<AuthContextType | undefined>(
+  undefined
+);
+
+export const AuthContextProvider = ({ children }: ContextProps) => {
+  const [user, setUser] = useState<User | null>(null);
+  const { addNotification } = useNotificationContext();
+
+  const login = useCallback(
+    (email: string) => {
+      setUser({ id: 1, name: "홍길동", email });
+      addNotification("성공적으로 로그인되었습니다", "success");
+    },
+    [setUser, addNotification]
+  );
+
+  const logout = useCallback(() => {
+    setUser(null);
+    addNotification("로그아웃되었습니다", "info");
+  }, [setUser, addNotification]);
+
+  const authCtxValue: AuthContextType = useMemo(
+    () => ({
+      user,
+      login,
+      logout,
+    }),
+    [user, login, logout]
+  );
+
+  return (
+    <AuthContext.Provider value={authCtxValue}>{children}</AuthContext.Provider>
+  );
+};
