--- src/context/NotificationContext.ts
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-import { NotificationContextType } from "../type/type";
-
-export const NotificationContext = createContext<
-  NotificationContextType | undefined
->(undefined);
