+++ src/context/ThemeContext.tsx
@@ -1,4 +1,5 @@
-import { createContext, useCallback, useMemo, useState } from "react";
+import { createContext, useState } from "react";
+import { useCallback, useMemo } from "../@lib";
 import { ThemeContextType, ContextProps } from "../type/type";
 
 export const ThemeContext = createContext<ThemeContextType | undefined>(
